server-python/requirements.txt
fastapi
uvicorn
chromadb
python-dotenv
requests
pymupdf